Move type-translator to type_translator.
... matching the common file name pattern in tsickle.
