+++ src/module_type_translator.ts
@@ -16,7 +16,7 @@ import * as googmodule from './googmodule';
 import * as jsdoc from './jsdoc';
 import {AnnotatorHost, isAmbient} from './jsdoc_transformer';
 import {createSingleQuoteStringLiteral, debugWarn, getIdentifierText, hasModifierFlag} from './transformer_util';
-import * as typeTranslator from './type-translator';
+import * as typeTranslator from './type_translator';
 import * as ts from './typescript';
 
 /**
